# Application/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "APIs - Desigualdade Conceitual Tecnológica",
        "description": "",
        "version": "1.0"
    },
    "produces": ["application/json"],
    "host": "https://dct.azurewebsites.net",
    "basePath": "/api",
    "paths": {
        "/user": {
            "get": {                        
                "operationId": "userLogin",
                "tags": ["User"],
                "description": "Login de Usuários",
                "parameters": [
                    {
                        "name" : "email",
                        "description": "Email do Usuário",
                        "required": true,
                        "schema": { "type": "string" }
                    },
                    {
                        "name" : "senha",
                        "description": "Senha do Usuário",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200" : { 
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }         
                    },
                    "400" : { "description": "Bad Request" },
                    "404" : { "description": "Not Found" }
                }
            },
            "put": {
                "operationId": "userUpdate",
                "tags": ["User"],
                "description": "Atualização de Usuários",
                "parameters": [

                ],
                "responses": {
                    "200" : { "description": "OK" },
                    "400" : { "description": "Bad Request" }
                }
            },
            "post": {
                "operationId": "userCreate",
                "tags": ["User"],
                "description": "Criação de Usuários",
                "parameters": [

                ],
                "responses": {
                    "201" : { "description": "Created" },
                    "400" : { "description": "Bad Request" }
                }
            },
            "delete": {
                "operationId": "userDelete",
                "tags": ["User"],
                "description": "Criação de Usuários",
                "parameters": [

                ],
                "responses": {
                    "204" : { "description": "No Content" },
                    "400" : { "description": "Bad Request" },
                    "404" : { "description": "Not Found" }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "_id": { "type": "string" }                    
                }
            }
        }
    }
}